# docs/swagger.yaml
basePath: /
definitions:
  application.KpiDto:
    properties:
      avgAge:
        type: number
      predomSpec:
        items:
          type: string
        type: array
      stdDev:
        type: number
    type: object
  application.PetDto:
    properties:
      age:
        example: 2
        type: integer
      dob:
        example: "2020-05-01T00:00:00Z"
        type: string
      gender:
        example: male
        type: string
      name:
        example: luke
        type: string
      species:
        example: dog
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Create pets and obtain stats from the pet DB
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Pet API
  version: "1.0"
paths:
  /CreatePet:
    post:
      consumes:
      - application/json
      description: Create a pet providing a pet model
      parameters:
      - description: pet
        in: body
        name: pet
        required: true
        schema:
          $ref: '#/definitions/application.PetDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/application.PetDto'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a pet
      tags:
      - PetController
  /Kpis:
    get:
      consumes:
      - application/json
      description: Get predominant species in the storage. Get average age and std
        deviation for the provided "species" parameter
      parameters:
      - description: species to get average age and std deviation from
        in: query
        name: species
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/application.KpiDto'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get pets Kpi
      tags:
      - PetController
  /ListPets:
    get:
      consumes:
      - application/json
      description: List all pets in the storage
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/application.PetDto'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List all pets
      tags:
      - PetController
swagger: "2.0"